[terraform] Use separate EBS volume for monitoring data
Create an EBS volume and mount at `/data` for monitoring data. This will
allow monitoring data to persist across instance restarts. We can also
snapshot the volume before destroying a workspace, and initialise a new
volume from a previous snapshot.

Also take the churn opportunity to rename resources from "prometheus" to
"monitoring".

Test Plan: Created mgorven workspace. Tainted monitoring instance to
replace it, monitoring data was not lost.

+++ terraform/templates/ec2_user_data.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
 
 if [ -e /dev/nvme1n1 ]; then
-	mkfs.ext4 /dev/nvme1n1
+	if ! file -s /dev/nvme1n1 | grep -q filesystem; then
+		mkfs.ext4 /dev/nvme1n1
+	fi
+
 	cat >> /etc/fstab <<-EOF
 	/dev/nvme1n1  /data  ext4  defaults,noatime  0  2
 	EOF
+
 	mkdir /data
 	mount /data
 fi
